requirements.txt
django
djangorestframework
djangorestframework
django-cors-headers
django-filter
mysql